pyproject.toml
[tool.poetry]
name = "fuser"
version = "0.1.0"
description = ""
authors = ["Alexandr Bobrov <alexbobrow@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg = {extras = ["binary"], version = "^3.2.6"}
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
